enhance(apps/frontend-manage): allow to choose language of live quiz embedding links
