+++ apps/docs/package.json
@@ -23,7 +23,7 @@
     "@tailwindcss/postcss": "~4.1.11",
     "@tailwindcss/typography": "~0.5.16",
     "@types/react": "^19.1.8",
-    "@uzh-bf/design-system": "4.0.2",
+    "@uzh-bf/design-system": "4.0.3",
     "autoprefixer": "~10.4.20",
     "cross-env": "~7.0.3",
     "docusaurus-plugin-matomo": "~0.0.8",
