+++ apps/frontend-manage/package.json
@@ -25,7 +25,7 @@
     "@socialgouv/matomo-next": "1.9.1",
     "@tanstack/react-table": "8.20.5",
     "@uidotdev/usehooks": "2.4.1",
-    "@uzh-bf/design-system": "4.0.2",
+    "@uzh-bf/design-system": "4.0.3",
     "dayjs": "1.11.13",
     "deepmerge": "4.3.1",
     "formik": "2.4.6",
