+++ apps/frontend-manage/src/components/liveQuiz/EmbeddingModal.tsx
@@ -1,54 +1,11 @@
 import { useQuery } from '@apollo/client'
-import { faClipboard } from '@fortawesome/free-regular-svg-icons'
 import { GetLiveQuizEmbeddingInfoDocument } from '@klicker-uzh/graphql/dist/ops'
-import { Button, Modal, Switch, toast } from '@uzh-bf/design-system'
+import { routing } from '@klicker-uzh/i18n'
+import { Modal, Select, Switch } from '@uzh-bf/design-system'
 import { useTranslations } from 'next-intl'
-import Link from 'next/link'
+import { useRouter } from 'next/router'
 import { useState } from 'react'
-
-function HMACLink({
-  quizId,
-  hmac,
-  params,
-  identifier,
-}: {
-  quizId: string
-  hmac: string
-  params: string
-  identifier: string
-}) {
-  const t = useTranslations()
-  const link = `${
-    process.env.NEXT_PUBLIC_MANAGE_URL
-  }/quizzes/${quizId}/evaluation?hmac=${hmac}${params ? `&${params}` : ''}`
-
-  return (
-    <div className="bg-accent flex max-w-full flex-row items-center justify-between gap-3 rounded px-2 py-1">
-      <Link href={link} legacyBehavior passHref>
-        <a
-          data-cy={`open-embedding-link-${identifier}`}
-          className="max-w-[calc(100%-3.5rem)] break-words text-sm"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          {link}
-        </a>
-      </Link>
-      <Button
-        onClick={() => {
-          navigator?.clipboard?.writeText(link)
-          toast({
-            type: 'success',
-            message: t('manage.liveQuizzes.embeddingLinkCopied'),
-          })
-        }}
-        data={{ cy: `copy-embed-link-live-quiz-${quizId}` }}
-      >
-        <Button.Icon withoutLabel icon={faClipboard} />
-      </Button>
-    </div>
-  )
-}
+import HMACLink from './HMACLink'
 
 function EmbeddingModal({
   onClose,
@@ -60,10 +17,20 @@ function EmbeddingModal({
   const t = useTranslations()
   const [showSolution, setShowSolution] = useState(false)
   const [showExplanation, setShowExplanation] = useState(false)
+  const router = useRouter()
   const { data, loading } = useQuery(GetLiveQuizEmbeddingInfoDocument, {
     variables: { id: quizId },
     fetchPolicy: 'cache-and-network',
   })
+
+  // language state for links
+  type LocaleType = (typeof routing.locales)[number]
+  const [language, setLanguage] = useState(
+    router.locale &&
+      (routing.locales as readonly string[]).includes(router.locale)
+      ? (router.locale as LocaleType)
+      : 'en'
+  )
 
   return (
     <Modal
@@ -84,7 +51,7 @@ function EmbeddingModal({
           checked={showSolution}
           onCheckedChange={(val) => setShowSolution(val)}
         />
-        <div className="pl-15 mb-3 text-sm">
+        <div className="pl-13 mb-3 text-sm">
           {t('manage.evaluation.showSolutionInfo')}
         </div>
         <Switch
@@ -93,8 +60,19 @@ function EmbeddingModal({
           checked={showExplanation}
           onCheckedChange={(val) => setShowExplanation(val)}
         />
-        <div className="pl-15 text-sm">
+        <div className="pl-13 mb-3 text-sm">
           {t('manage.evaluation.showExplanationInfo')}
+        </div>
+        <div className="pl-13">
+          <div className="font-bold">{t('shared.generic.language')}</div>
+          <Select
+            value={language}
+            onChange={(newValue) => setLanguage(newValue as LocaleType)}
+            items={[
+              { label: t('shared.generic.english'), value: 'en' },
+              { label: t('shared.generic.german'), value: 'de' },
+            ]}
+          />
         </div>
       </div>
 
@@ -106,6 +84,7 @@ function EmbeddingModal({
             hmac={data?.getLiveQuizEmbeddingInfo?.hmac ?? ''}
             params=""
             identifier="generic-evaluation"
+            language={language}
           />
         </div>
         {data?.getLiveQuizEmbeddingInfo?.instances?.map((instance, ix) => {
@@ -119,6 +98,7 @@ function EmbeddingModal({
                 hmac={data?.getLiveQuizEmbeddingInfo?.hmac ?? ''}
                 params={`questionIx=${ix}&hideControls=true&showSolution=${showSolution}&showExplanation=${showExplanation}`}
                 identifier={`question-${ix}`}
+                language={language}
               />
             </div>
           )
@@ -130,8 +110,9 @@ function EmbeddingModal({
           <HMACLink
             quizId={quizId}
             hmac={data?.getLiveQuizEmbeddingInfo?.hmac ?? ''}
-            params={`leaderboard=true&hideControls=true`}
+            params="leaderboard=true&hideControls=true"
             identifier="leaderboard"
+            language={language}
           />
         </div>
       </div>
