+++ apps/frontend-pwa/package.json
@@ -25,7 +25,7 @@
     "@klicker-uzh/util": "workspace:*",
     "@socialgouv/matomo-next": "1.9.1",
     "@uidotdev/usehooks": "2.4.1",
-    "@uzh-bf/design-system": "4.0.2",
+    "@uzh-bf/design-system": "4.0.3",
     "body-parser": "1.20.3",
     "dayjs": "1.11.13",
     "deepmerge": "4.3.1",
