+++ apps/office-addin/package.json
@@ -3,7 +3,7 @@
   "version": "3.3.0-alpha.77",
   "license": "AGPL-3.0",
   "dependencies": {
-    "@uzh-bf/design-system": "4.0.2",
+    "@uzh-bf/design-system": "4.0.3",
     "core-js": "3.30.2",
     "es6-promise": "4.2.8",
     "formik": "2.4.6",
