+++ packages/markdown/package.json
@@ -40,7 +40,7 @@
     "@fortawesome/free-regular-svg-icons": "^6.7.2",
     "@fortawesome/free-solid-svg-icons": "^6.7.2",
     "@fortawesome/react-fontawesome": "^0.2.2",
-    "@uzh-bf/design-system": "4.0.2",
+    "@uzh-bf/design-system": "4.0.3",
     "next": "^15.3.4",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
