+++ packages/shared-components/package.json
@@ -33,7 +33,7 @@
     "@klicker-uzh/graphql": "workspace:*",
     "@klicker-uzh/markdown": "workspace:*",
     "@tailwindcss/postcss": "^4.1.11",
-    "@uzh-bf/design-system": "4.0.2",
+    "@uzh-bf/design-system": "4.0.3",
     "dayjs": "^1.11.13",
     "formik": "^2.4.6",
     "js-cookie": "^3.0.5",
