+++ packages/transactional/package.json
@@ -8,7 +8,7 @@
     "@react-email/components": "~0.0.25",
     "@types/node": "^20.16.1",
     "@types/react": "^19.1.8",
-    "@uzh-bf/design-system": "4.0.2",
+    "@uzh-bf/design-system": "4.0.3",
     "lucide-react": "~0.522.0",
     "react": "~19.1.0",
     "react-email": "~3.0.1",
